// types/earningsRevenue.ts
export interface EarningsRevenueInterface {
  Date: string
  Symbol: string
  Name: string
  Actual: string
  ActualValue: number
  Forecast: string
  ForecastValue: number
  Previous: string
  PreviousValue: number
  Revenue: string
  RevenueValue: number
  RevenueForecast: string
  RevenueForecastValue: number
  RevenuePrevious: string
  RevenuePreviousValue: number
  MarketCapUSD: number
  FiscalTag: string
  FiscalReference: string
  CalendarReference: string
  Country: string
  Currency: string
  Importance: number
  Session: number
  marketRelease: string
  LastUpdate: string
}